Refactor `src/core/vector.rs` into a directory-based module
- Split `src/core/vector.rs` into `types.rs`, `constants.rs`, `metric.rs`, `validation.rs`, `simd.rs`, `ops.rs`, `sparse.rs`, and `tests.rs`.
- Preserve public API via re-exports in `src/core/vector/mod.rs`.
- Expose internal helpers like `dot_and_magnitudes` as `pub(crate)` for use within the module.
- Fix imports in tests.
- Improve maintainability by reducing file size from ~5300 lines to multiple focused files.
